transfer (view)models and interfaces into separate project, rename namespaces
