+++ Common/GbWebApp.Domain/ViewModels/PaginationViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace GbWebApp.ViewModels
+namespace GbWebApp.Domain.ViewModels
 {
     public class PaginationViewModel
     {
