+++ GbAsp2.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GbWebApp.DAL", "Services\Gb
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GbWebApp", "UI\GbWebApp\GbWebApp.csproj", "{543BC2E1-521A-4121-9E21-CC648A983E3C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GbWebApp.Interfaces", "Services\GbWebApp.Interfaces\GbWebApp.Interfaces.csproj", "{14E10FA4-D76B-49A4-A4DA-B624658658A4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{543BC2E1-521A-4121-9E21-CC648A983E3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{543BC2E1-521A-4121-9E21-CC648A983E3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{543BC2E1-521A-4121-9E21-CC648A983E3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14E10FA4-D76B-49A4-A4DA-B624658658A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14E10FA4-D76B-49A4-A4DA-B624658658A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14E10FA4-D76B-49A4-A4DA-B624658658A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14E10FA4-D76B-49A4-A4DA-B624658658A4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -41,6 +47,7 @@ Global
 		{7B23AA62-AF66-45B5-BCDB-0419F28415B8} = {18461E5A-1239-4989-95C7-88286FCA6E10}
 		{EC94F462-1B15-4A3F-BB1A-BA67194B0629} = {52C8CCD4-C112-44F9-B36C-7F74C3B935DD}
 		{543BC2E1-521A-4121-9E21-CC648A983E3C} = {58D9F3FA-201A-42CA-A8FD-7AD034A5E555}
+		{14E10FA4-D76B-49A4-A4DA-B624658658A4} = {52C8CCD4-C112-44F9-B36C-7F74C3B935DD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1FB38A22-85FF-415D-9199-682C0986E2B3}
