+++ Services/GbWebApp.Interfaces/GbWebApp.Interfaces.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common\GbWebApp.Domain\GbWebApp.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
