+++ UI/GbWebApp/Areas/Admin/Controllers/ProductsController.cs
@@ -1,12 +1,12 @@
-﻿using GbWebApp.Infrastructure.Interfaces;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using GbWebApp.Domain.Entities.Identity;
 using GbWebApp.Infrastructure.Mappers;
 using Microsoft.EntityFrameworkCore;
 using GbWebApp.Domain.Entities;
 using Microsoft.AspNetCore.Mvc;
-using GbWebApp.ViewModels;
 using System;
+using GbWebApp.Domain.ViewModels;
+using GbWebApp.Interfaces.Services;
 
 namespace GbWebApp.Areas.Admin.Controllers
 {
