+++ UI/GbWebApp/Areas/Admin/Views/Products/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProductViewModel
+﻿@model GbWebApp.Domain.ViewModels.ProductViewModel
 
 @{ ViewBag.Title = "Admin - Product Edit"; }
