+++ UI/GbWebApp/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,5 +1,4 @@
 ﻿@using GbWebApp.Domain.Entities
-@using GbWebApp.ViewModels
 @using GbWebApp
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
