+++ UI/GbWebApp/Components/BrandsViewComponent.cs
@@ -1,10 +1,10 @@
-﻿using GbWebApp.Infrastructure.Interfaces;
-using GbWebApp.ViewModels;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using GbWebApp.Domain.ViewModels;
+using GbWebApp.Interfaces.Services;
 
 namespace GbWebApp.Components
 {
