+++ UI/GbWebApp/Infrastructure/Services/InDB/InDbAnyEntity.cs
@@ -1,8 +1,8 @@
 ﻿using GbWebApp.DAL.Context;
 using GbWebApp.Domain.Entities.Base.Interfaces;
-using GbWebApp.Infrastructure.Interfaces;
 using System;
 using System.Linq;
+using GbWebApp.Interfaces.Services;
 using Microsoft.EntityFrameworkCore;
 
 namespace GbWebApp.Infrastructure.Services.InDB
