+++ UI/GbWebApp/Startup.cs
@@ -2,7 +2,6 @@ using GbWebApp.Infrastructure.Services.InCookies;
 using Microsoft.Extensions.DependencyInjection;
 using GbWebApp.Infrastructure.Services.InDB;
 using Microsoft.Extensions.Configuration;
-using GbWebApp.Infrastructure.Interfaces;
 using GbWebApp.Domain.Entities.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Identity;
@@ -11,6 +10,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 using GbWebApp.DAL.Context;
 using GbWebApp.Data;
+using GbWebApp.Interfaces.Services;
 
 namespace GbWebApp
 {
