+++ UI/GbWebApp/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model LoginViewModel
+﻿@model GbWebApp.Domain.ViewModels.LoginViewModel
 @{ ViewBag.Title = "Login"; }
 
 <section id="form">
