+++ UI/GbWebApp/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model RegisterViewModel
+﻿@model GbWebApp.Domain.ViewModels.RegisterViewModel
 @{ ViewBag.Title = "Register"; }
 
 <section id="form">
