+++ UI/GbWebApp/Views/Employees/Emp_Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeViewModel
+﻿@model GbWebApp.Domain.ViewModels.EmployeeViewModel
 
 @{ ViewBag.Title = "Edit "; }
