+++ UI/GbWebApp/Views/Employees/Emp_New.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeViewModel
+﻿@model GbWebApp.Domain.ViewModels.EmployeeViewModel
 
 @{ ViewBag.Title = "New... "; }
