+++ UI/GbWebApp/Views/Home/Index.cshtml
@@ -1,5 +1,5 @@
-﻿@using GbWebApp.Infrastructure.Interfaces
-@using GbWebApp.Infrastructure.Mappers
+﻿@using GbWebApp.Infrastructure.Mappers
+@using GbWebApp.Interfaces.Services
 @inject IProductData productData
 
 @{
