+++ UI/GbWebApp/Views/Home/Partial/_Featured.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ProductViewModel>
+﻿@model IEnumerable<GbWebApp.Domain.ViewModels.ProductViewModel>
 
 <div class="features_items">
     <h2 class="title text-center">Featured Items</h2>
