+++ UI/GbWebApp/Views/Shared/Components/Brands/ComboId.cshtml
@@ -1,4 +1,4 @@
-﻿@model (int, IEnumerable<BrandsViewModel>)
+﻿@model (int, IEnumerable<GbWebApp.Domain.ViewModels.BrandsViewModel>)
 
 @*<select required asp-for="BrandId" class="form-control">*@
 @*<option value="" disabled>Please select...</option>*@
