+++ UI/GbWebApp/Views/Shared/Components/Brands/ComboNew.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<BrandsViewModel>
+﻿@model IEnumerable<GbWebApp.Domain.ViewModels.BrandsViewModel>
 
 @*<select required class="form-control">*@
 <option value="" disabled selected>Please select...</option>
