+++ UI/GbWebApp/Views/Shared/Components/Brands/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<BrandsViewModel>
+﻿@model IEnumerable<GbWebApp.Domain.ViewModels.BrandsViewModel>
 <div class="brands_products">
     <h2>Brands</h2>
     <div class="brands-name">
