+++ UI/GbWebApp/Views/Shared/Components/Sections/ComboId.cshtml
@@ -1,4 +1,4 @@
-﻿@model (int, IEnumerable<SectionViewModel>)
+﻿@model (int, IEnumerable<GbWebApp.Domain.ViewModels.SectionViewModel>)
 
 <select required name="SectionId" class="form-control">@*"asp-for" option doesn't work, but "name" works... however, you must be very carefully with it!*@
     @*<option value="" disabled>Please select...</option>*@
