+++ UI/GbWebApp/Views/Shared/Components/Sections/ComboNew.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<SectionViewModel>
+﻿@model IEnumerable<GbWebApp.Domain.ViewModels.SectionViewModel>
 
 <select required name="SectionId" class="form-control">@*"asp-for" option doesn't work, but "name" works... however, you must be very carefully with it!*@
     <option value="" disabled selected>Please select...</option>
