+++ UI/GbWebApp/Views/Shop/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model ShopViewModel
+﻿@model GbWebApp.Domain.ViewModels.ShopViewModel
 
 @{ ViewBag.Title = "Shop "; }
