+++ UI/GbWebApp/Views/Shop/Partial/_CartItems.cshtml
@@ -1,4 +1,4 @@
-﻿@model CartViewModel
+﻿@model GbWebApp.Domain.ViewModels.CartViewModel
 
 <section id="cart_items">
     <div class="container">
