+++ UI/GbWebApp/Views/Shop/Partial/_DoAction.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrderViewModel
+﻿@model GbWebApp.Domain.ViewModels.OrderViewModel
 
 <section id="do_action">
     <div class="container">
