+++ UI/GbWebApp/Views/Shop/Partial/_ProductDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProductViewModel
+﻿@model GbWebApp.Domain.ViewModels.ProductViewModel
 
 <div class="product-details">
     <div class="col-sm-5">
