+++ UI/GbWebApp/Views/Shop/ShopCart.cshtml
@@ -1,4 +1,4 @@
-﻿@model (CartViewModel, OrderViewModel)
+﻿@model (GbWebApp.Domain.ViewModels.CartViewModel, GbWebApp.Domain.ViewModels.OrderViewModel)
 @{ ViewBag.Title = "Cart "; }
 
 <section>
