+++ UI/GbWebApp/Views/Shop/ShopDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProductViewModel
+﻿@model GbWebApp.Domain.ViewModels.ProductViewModel
 @{ ViewBag.Title = $"Product '{Model.Name}' Details "; }
 
 <section>
