+++ UI/GbWebApp/Views/Users/Orders.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<UserOrderViewModel>
+﻿@model IEnumerable<GbWebApp.Domain.ViewModels.UserOrderViewModel>
 
 @{
     ViewBag.Title = "User Orders ";
