+++ UI/GbWebApp/Views/_ViewImports.cshtml
@@ -1,6 +1,5 @@
 ﻿@using GbWebApp
 @using GbWebApp.Domain.Entities
-@using GbWebApp.ViewModels
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, GbWebApp
